requirements.txt
openai
streamlit
python-dotenv
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
pdfminer.six
python-docx